[Feat] BaseInterceptor로 클래스 분리
